cmake: Add most of standalone/ to the build system
The all target now builds more of the standalone stuff and the install
target installs everything that was built. If you want to exclude the
standalone targets, you can build just the hiddenstar target, but the
only way to avoid installing the html file right now is to delete it.

The standalone starmap was excluded from this, since it has bit-rotted.
Normally I'd just remove it from the project, but I was advised not to.
If you want to work on fixing it, you'll have to compile it without
cmake for now, or add the target yourself.